se agregan validaciones y test al servicio de calcularRentaAuto

+++ java-arquitectura-hexagonal/microservicio/dominio/src/main/java/com/ceiba/tarifa/modelo/dto/DtoTarifa.java
@@ -12,5 +12,5 @@ import java.math.BigDecimal;
 public class DtoTarifa {
     private Long id;
 	private String tipoCombustible;
-	private BigDecimal porcentaje;
+	private Double porcentaje;
 }
