+++ java-arquitectura-hexagonal/microservicio/infraestructura/src/main/java/com/ceiba/tarifa/adaptador/dao/MapeoTarifa.java
@@ -1,9 +1,7 @@
 package com.ceiba.tarifa.adaptador.dao;
 
-import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-
 
 import com.ceiba.infraestructura.jdbc.MapperResult;
 import com.ceiba.tarifa.modelo.dto.DtoTarifa;
@@ -11,14 +9,14 @@ import org.springframework.jdbc.core.RowMapper;
 
 public class MapeoTarifa implements RowMapper<DtoTarifa>, MapperResult {
 
-    @Override
-    public DtoTarifa mapRow(ResultSet resultSet, int rowNum) throws SQLException {
+	@Override
+	public DtoTarifa mapRow(ResultSet resultSet, int rowNum) throws SQLException {
 
-        Long id = resultSet.getLong("id");
-        String tipoCombustible = resultSet.getString("tipoCombustible");
-        BigDecimal porcentaje = resultSet.getBigDecimal("porcentaje");
+		Long id = resultSet.getLong("id");
+		String tipoCombustible = resultSet.getString("tipoCombustible");
+		Double porcentaje = resultSet.getDouble("porcentaje");
 
-        return new DtoTarifa(id,tipoCombustible,porcentaje);
-    }
+		return new DtoTarifa(id, tipoCombustible, porcentaje);
+	}
 
 }
